+++ lib/src/core/exception.dart
@@ -1,9 +1,35 @@
+/// Exception thrown when an the server returns an error code to an rpc request.
+class WebThreeRPCError implements Exception {
+  final int code;
+  final String message;
+  final dynamic data;
+
+  const WebThreeRPCError(this.code, this.message, this.data);
+
+  @override
+  String toString() {
+    return 'RPCError: got code $code with msg \"$message\".';
+  }
+}
+
 class EthereumUserRejected implements Exception {
+  final int code;
+  final String message;
+  final dynamic data;
+
+  EthereumUserRejected(this.code, this.message, this.data);
+
   @override
   String toString() => 'EthereumUserRejected: User rejected the request';
 }
 
 class EthereumChainSwitchNotSupported implements Exception {
+  final int code;
+  final String message;
+  final dynamic data;
+
+  EthereumChainSwitchNotSupported(this.code, this.message, this.data);
+
   @override
   String toString() =>
       'EthereumChainSwitchNotSupported: User rejected the request';