+++ lib/src/core/exception_utils_io.dart
@@ -6,7 +6,7 @@ class ExceptionUtils {
   /// Analyze Exception to throw a typed exception
   static void analyzeException(Object e) {
     if (e.toString().toLowerCase().contains('user rejected the request')) {
-      throw EthereumUserRejected();
+      throw EthereumUserRejected(9999, 'native exception', '');
     } else {
       throw UnimplementedError(e.toString());
     }