+++ lib/webthree.dart
@@ -18,8 +18,8 @@ import 'crypto.dart';
 import 'json_rpc.dart';
 import 'src/core/block_information.dart';
 import 'src/core/block_number.dart';
+import 'src/core/exception.dart';
 import 'src/core/sync_information.dart';
-
 import 'src/utils/rlp.dart' as rlp;
 import 'src/utils/typed_data.dart';
 
@@ -30,11 +30,10 @@ export 'src/core/block_number.dart';
 export 'src/core/exception.dart';
 export 'src/core/sync_information.dart';
 
-part 'src/core/ether_unit.dart';
-part 'src/core/ether_amount.dart';
-part 'src/core/eip1559_information.dart';
-
 part 'src/core/client.dart';
+part 'src/core/eip1559_information.dart';
+part 'src/core/ether_amount.dart';
+part 'src/core/ether_unit.dart';
 part 'src/core/filters.dart';
 part 'src/core/transaction.dart';
 part 'src/core/transaction_information.dart';